+++ src/main/java/calculator/Operation.java
@@ -1,0 +1,6 @@
+package calculator;
+
+@FunctionalInterface
+public interface Operation {
+    int calculate(int first, int second);
+}
