+++ src/main/java/racingcar/README.md
@@ -1,0 +1,24 @@
+
+##### 구현기능 목록
+- RacingCar 도메인 기능 
+- 사용자 질의 응답 기능
+- 랜덤명령 생성 테스트
+- 참가 차량 댓수에 맞는 RacingCar 인스턴스 생성
+- 각 참가차량에 랜덤명령을 전달 테스트
+
+- 각 자동차에 이름을 부여할 수 있다. 자동차 이름은 5자를 초과할 수 없다.
+- 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
+- 자동차 이름은 쉼표(,)를 기준으로 구분한다.
+- 자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.
+
+##### Step3 
+ - RaceGameContract 인터페이스 : 이너 Presenter, View 인터페이스를 선언함으로써 특정 도메인에서만 사용되도록 제한한다. 
+ - Presenter : view와 도메인 모델의 중개역할을 담당한다. 데이터의 흐름을 제어하는 역할 (비즈니스 로직은 절대 제외할것)
+ - Commander : 게임규칙을 표현하는 인터페이스 
+ - RacingCar : 참가차량의 상태를 표현하는 클래스
+ - RacingCars : 모든 참가차량을 표현하는 클래스 
+ - RaceRound : Round 상태를 표현하는 클래스 
+ - RaceGame : Race 게임 start와 end를 표현하는 클래스 
+ 
+ 
+
