+++ src/main/java/racingcar/RaceGameContract.java
@@ -1,0 +1,14 @@
+package racingcar;
+
+import racingcar.model.RaceGame;
+
+public interface RaceGameContract {
+
+    interface View {
+        void renderView(RaceGame raceGame);
+    }
+
+    interface Presenter {
+        void gameStart();
+    }
+}
