+++ src/main/java/racingcar/exception/NotFinishRaceGameException.java
@@ -1,0 +1,10 @@
+package racingcar.exception;
+
+public class NotFinishRaceGameException extends RuntimeException {
+
+    private static final String ERROR_MESSAGE = " 라운드가 아직 남았어요";
+
+    public NotFinishRaceGameException(int round) {
+        super(round + ERROR_MESSAGE);
+    }
+}
