+++ src/main/java/racingcar/model/Commander.java
@@ -1,0 +1,5 @@
+package racingcar.model;
+
+public interface Commander {
+    boolean moveForward();
+}
