+++ src/main/java/racingcar/present/Presenter.java
@@ -1,0 +1,23 @@
+package racingcar.present;
+
+import racingcar.RaceGameContract;
+import racingcar.model.RaceGame;
+
+public class Presenter implements RaceGameContract.Presenter {
+
+    private final RaceGameContract.View view;
+    private final RaceGame raceGame;
+
+    public Presenter(RaceGameContract.View view, RaceGame raceGame) {
+        this.view = view;
+        this.raceGame = raceGame;
+    }
+
+    @Override
+    public void gameStart() {
+        raceGame.start();
+        view.renderView(raceGame);
+    }
+
+
+}
