--- src/main/java/step2/Operation.java
@@ -1,6 +1,0 @@
-package step2;
-
-@FunctionalInterface
-public interface Operation {
-    int calculate(int first, int second);
-}
