--- src/main/java/step3/RaceGameContract.java
@@ -1,14 +1,0 @@
-package step3;
-
-import step3.model.RaceGame;
-
-public interface RaceGameContract {
-
-    interface View {
-        void renderView(RaceGame raceGame);
-    }
-
-    interface Presenter {
-        void gameStart();
-    }
-}
