--- src/main/java/step3/exception/NotFinishRaceGameException.java
@@ -1,10 +1,0 @@
-package step3.exception;
-
-public class NotFinishRaceGameException extends RuntimeException {
-
-    private static final String ERROR_MESSAGE = " 라운드가 아직 남았어요";
-
-    public NotFinishRaceGameException(int round) {
-        super(round + ERROR_MESSAGE);
-    }
-}
