--- src/main/java/step3/model/Commander.java
@@ -1,5 +1,0 @@
-package step3.model;
-
-public interface Commander {
-    boolean moveForward();
-}
