--- src/main/java/step3/model/RacingCar.java
@@ -1,32 +1,0 @@
-package step3.model;
-
-public class RacingCar {
-
-    private int targetPoint;
-    private int currentPosition;
-    private Commander commander;
-
-    public RacingCar(Commander commander) {
-        this.commander = commander;
-        this.targetPoint++;
-    }
-
-    public void executeMoveTrack() {
-        if(commander.moveForward()){
-            targetPoint ++;
-        }
-    }
-
-    public int movePosition() {
-        if(isReachMoveTargetPosition()) {
-            return targetPoint;
-        }
-        currentPosition++;
-        return currentPosition;
-    }
-
-
-    private boolean isReachMoveTargetPosition() {
-        return currentPosition == targetPoint ? true : false;
-    }
-}
